azure_functions/agents/Twitter/requirements.txt
azure-functions
requests==2.32.3
beautifulsoup4==4.13.4
python-dotenv==1.1.1
selenium==4.24.0
webdriver-manager==4.0.2

